I've completed the refactoring you requested. Here's a summary of the changes:
Refactor: Consolidate jules-feedbin-client into google-jules and restore jetbrains-junie

This work addresses several repository cleanup and consolidation tasks:

1.  **Restore `feedbin-api/jetbrains-junie/`**:
    The `feedbin-api/jetbrains-junie/` directory has been restored to its state from the `main` branch.

2.  **Consolidate `jules-feedbin-client` into `google-jules`**:
    The Feedbin client library previously in `feedbin-api/jules-feedbin-client/` has been merged into `feedbin-api/google-jules/`.
    - All Go source files are now located in `feedbin-api/google-jules/feedbinapi/`.
    - The `go.mod`, `go.sum`, `README.md`, and `examples/` directory from `jules-feedbin-client` are now in `feedbin-api/google-jules/`.
    - Package declarations for the library files have been updated to `package feedbinapi`.
    - Import paths in the example code (`feedbin-api/google-jules/examples/main.go`) have been adjusted accordingly.
    - The `feedbin-api/jules-feedbin-client/` directory has been removed.

3.  **Build and Dependency Management**:
    - I ensured dependencies are correct within `feedbin-api/google-jules/`.
    - I resolved "imported and not used" build errors in the `feedbin-api/google-jules/feedbinapi/` package.
    - The `feedbin-api/google-jules/` project now builds successfully and its checks pass.
    - The restored `feedbin-api/jetbrains-junie/` project also builds successfully.

All changes related to the `jules-feedbin-client` merge are now contained within the `feedbin-api/google-jules/` directory structure.
