+++ feedbin-api/google-jules/feedbinapi/client_test.go
@@ -1,4 +1,4 @@
-package feedbin
+package feedbinapi
 
 import (
 	"fmt"
