+++ feedbin-api/google-jules/feedbinapi/entries.go
@@ -1,4 +1,4 @@
-package feedbin
+package feedbinapi
 
 import (
 	"fmt"
