+++ feedbin-api/google-jules/feedbinapi/extract.go
@@ -1,4 +1,4 @@
-package feedbin
+package feedbinapi
 
 import (
 	"encoding/json"
