+++ feedbin-api/google-jules/feedbinapi/pages.go
@@ -1,4 +1,4 @@
-package feedbin
+package feedbinapi
 
 import (
 	"fmt"
