+++ feedbin-api/google-jules/feedbinapi/unread_entries.go
@@ -1,9 +1,8 @@
-package feedbin
+package feedbinapi
 
 import (
 	"fmt"
 	"net/http"
-	"strings"
 
 	"github.com/google/go-querystring/query" // For potential future options
 )
