+++ feedbin-api/google-jules/go.mod
@@ -1,3 +1,5 @@
-module github.com/jules/feedbinapi
+module jules-feedbin-client
 
 go 1.22.2
+
+require github.com/google/go-querystring v1.1.0
