--- feedbin-api/jules-feedbin-client/go.mod
@@ -1,5 +1,0 @@
-module jules-feedbin-client
-
-go 1.22.2
-
-require github.com/google/go-querystring v1.1.0
